refactor: 상수화가 안된 부분 상수화, 필요하지 않은 모듈 불러오기 제거
